Code movement, added workflow execution print
